BUG: Correctly compute xcorr
Previously, the xcorr function provided assumed that both input time
series (i.e. a template waveform and a real-time trace) were demeaned.
However, this assumption is incorrect, since only the template waveform
data is correctly demeaned.

The implementation provided by this commit is based on a single-pass
algorithm of the Pearson correlation coefficient.

Make XCorr a method of Template.