MAINT/BUG: Use Core::SmartPointer
The design of Processor does not allow us to use std::unique_ptr since
unique ownership is not granted. Particularly, this is an issue when
(method) callbacks via boost::bind are used.

As a consequence, weaken both DetectorBuilder and TemplateBuilder i.e.
return a Core::SmartPointer instead of std::unique_ptr.

Usually, builders should return a std::unique_ptr to the product built.
However, in the application specific use case (due to callbacks etc.) it
is preferable to return a Core::SmartPointer, immediately.