BUG: Return the max coefficient
The Pearson correlation coefficient has a value between +1 and -1. A
value of +1 is total positive linear correlation, 0 is no linear
correlation, and -1 is total negative linear correlation.

Comparing absolute values might lead to a situation where the negative
linear correlation is greater than the positive linear correlation.
However, when comparing waveforms, we're not interested in negative
linear correlation, but in the positive linear correlation.