ENH: Allow any phase code to be configured
When configuring a `Detector` with a stream (by means of
`DetectorBuilder::set_stream()`) the sensor location specified by the
template stream identifier is compared with the waveform stream
identifier associated with the corresponding pick. This is critical
since certain picks are picked on projected/processed waveforms, and
thus the template waveform stream identifier does not match with the
pick's stream identifier, anymore.

Besides, any phase code is accepted provided by module users.

The implemenation follows the conclusion of the ongoing discussion of
issue #3.

Closes: #3