MAINT: Set default values
Use member default initializers in order to set default values for both
the gap tolerance and the gap threshold.

Note that not configuring the gap threshold (i.e. the threshold is set to zero)
requires every single gap to be handled. However, actually it would be nice to
have the gap threshold dependent on the sampling frequency in order to avoid
the situation that gaps smaller than 1/(sampling frequency) seconds are going
to be handled.