ENH: Avoid distortions when filtering template waveform
Since SeisComP filter implementations are implemented as recursive
filters, the data would contain distortions if filtered. In order to
avoid such distortions, apply the filter to a larger time window
(including the data) taking the margin into account just to trim the
data to the final time window desired. Alternatively, we could apply a
taper of appropriate length before filtering.