MAINT: Link against sqlite3 library
This approach requires maintaining a copy of the SQLite database interface
internally where the service is named `sqlite3_`. In contrast the the original
attempt i.e. loading the standard SeisComP SQLite database interface plugin
this is the approach recommended by gempa-jabe. That is, if the application
requires a peace of code, then link against it. Also, from the user perspective
this is the cleanest approach.

Note that the SQLite database interface implemented by the SeisComP core
framework is provided as a C++ module plugin library, and thus cannot be used
for linking.