ENH: Implement amplitude processor facilities
- provide `TimeWindowProcessor` interface
- provide `AmplitudeProcessor` and `ReducingAmplitudeProcessor`
  interfaces
- implement `RMSAmplitude` processor which is intended to compute the
  amplitude similar to `seisMatch`

Note that currently amplitude processing facilities do not implement
methods in order to compute the SNR.

Also note that this is an alternative implementation to
`Processing::AmplitudeProcessor`. It is intended to provide a simplified
interface to be used for processors with reducing facilities (i.e.
processors which are required to handle multiple streams
simultaneously).