ENH: Allow to configure limits
- Allow users to configure regression limits for those template family
members referencing a detector (i.e. those which are going to *produce*
new amplitudes and thus magnitudes).
- Allow users to provide limits by means of the `"lowerLimit"` and
"upperLimit" template family configuration parameter. Note that those
parameters can be provided as so called *stream configuration defaults*.
- Store the sensor location identifier within template family members.
This is required in order to correctly evaluate limits, after performing
the regression. (Note that storing the detector identifier might be
required, too. Though, this will be implemented in a subsequent
commit.)