ENH: Drop intermediate buffers
Do not buffer records. Instead, process records on-the-fly. This
approach should result in a significant speed up, since records are
processed as they are fed into the detector. Besides, data is not
required to be aligned, anymore.

Note that the gap interpolation is now handled by the detector waveform
processor, while previously it was handled by the ringbuffer operator.