BUG: Implement pick time based triggering
This is similar to the approach implemented by `scautopick`. Note that
the previous implementation which was record endtime was flawed. Due to
the fact that data is fed record-wise the trigger resulution was bound
to both the record size and the data sampling frequency (i.e. in case
the record size is fixed (e.g. MSEED)).