ENH: Compute `MRelative` based on NetworkMagnitude
Estimating `MRelative` based on the *NetworkMagnitude* (i.e. the
preferred event magnitude) by default. Optionally, `MRelative` still may
be estimated based on the StationMagnitude. Wether to use the
NetworkMagnitude or the StationMagnitude may be defined in the bindings
configuration (as a consequence, this parameter is configurable on
sensor location granularity).

Note that for estimating `MRelative` based on the NetworkMagnitude
requires both the Event and the Origin including the preferred event
magnitude to be available in the EventParameters data.

Besides, with this commit the sensor location identifiers referencing
StationMagnitudes (at `magnitude::Factory`) include the band and source
code.